PHPCS: Add a starting point for local PHPCS setup
Background: https://keepingtheirblogsgoing.wordpress.com/2018/11/24/proposal-to-add-default-phpcs-configuration-to-skeleton-repo/

Developers (clients, agencies, VIP), can run `composer install`, and have the same versions of PHPCS, standards (WPCS, VIPCS, PHPCompatibilityWP), and use the same PHPCS configuration (collection of rules and flags).

This not only gives developers a starting point in trying to adhere to a workflow that uses PHPCS locally (and not just relying on a VIP review), but it also exposes the configuration so everyone is working from the same setup and therefore get the same violations reported.